Combine DeployableMicroservicesList with MicroserviceNewScreen
