+++ Source/SelfService/Web/applications/microservice/microserviceNewScreen.tsx
@@ -1,15 +1,82 @@
 // Copyright (c) Dolittle. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 
-import { DeployableMicroservicesList } from './deployableMicroservicesList';
+import { useLocation, useNavigate } from 'react-router';
+
+import { Typography } from '@mui/material';
+
+import { SimpleCardGrid, SimpleCardProps } from '@dolittle/design-system';
 
 import { HttpResponseApplication } from '../../apis/solutions/application';
+
+import { DeployMicroservice } from './deployMicroservice/deployMicroservice';
+
+type DeployableMicroservices = SimpleCardProps & {
+    kind: string;
+};
 
 export type MicroserviceNewScreenProps = {
     application: HttpResponseApplication;
 };
 
-export const MicroserviceNewScreen = ({ application }: MicroserviceNewScreenProps) =>
-    <DeployableMicroservicesList application={application} />;
+export const MicroserviceNewScreen = ({ application }: MicroserviceNewScreenProps) => {
+    const navigate = useNavigate();
+    const location = useLocation();
+
+    const searchParams = new URLSearchParams(location.search);
+
+    const deployableMicroservices: DeployableMicroservices[] = [
+        {
+            kind: 'dolittle-microservice',
+            title: 'Base (default)',
+            description: 'Setup a container with the Dolittle runtime ready to consume your events.',
+            secondaryButton: {
+                label: 'Learn more',
+                href: 'https://dolittle.io/docs/platform/requirements/',
+                target: true,
+                ariaLabel: 'Learn more about the base microservice.',
+            },
+            primaryButton: {
+                label: 'Deploy',
+                onClick: () => onCreate('dolittle-microservice'),
+            },
+        },
+    ];
+
+    const kindViaParams = () => {
+        if (!searchParams.has('kind')) return '';
+        const kind = searchParams.get('kind') as string;
+        if (!kind || !deployableMicroservices.map(item => item.kind).includes(kind)) return '';
+
+        return kind;
+    };
+
+    const [microserviceTypeState, setMicroserviceTypeState] = useState(kindViaParams());
+
+    useEffect(() => {
+        setMicroserviceTypeState(kindViaParams());
+    }, [kindViaParams()]);
+
+    const onCreate = (kind: string) => {
+        searchParams.set('kind', kind);
+        navigate({ pathname: location.pathname, search: searchParams.toString() }, { replace: true });
+        setMicroserviceTypeState(kind);
+    };
+
+    if (microserviceTypeState === '') {
+        return (
+            <>
+                <Typography variant='h1' my={3}>Microservices</Typography>
+                <SimpleCardGrid simpleCardItems={deployableMicroservices} />
+            </>
+        );
+    }
+
+    return (
+        <>
+            {microserviceTypeState === 'dolittle-microservice' && <DeployMicroservice application={application} />}
+        </>
+    );
+};
